liver_grading_system/requirements.txt
flask==2.3.3
werkzeug==2.3.7
tensorflow==2.16.1
numpy>=1.24.0
pillow>=8.3.2
scikit-learn>=0.24.2
pandas>=1.3.3
matplotlib>=3.4.3 